# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LaunchReady API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GeneratePrdBody {
  problem: string;
  audience: string;
  success: string;
  /** @nullable */
  productName?: string | null;
}

export interface TaskItem {
  id: string;
  title: string;
  description: string;
  dependsOn: string[];
  status: string;
  effort: string;
  effortScore: number;
  isBlocked: boolean;
  blockedBy: string[];
  sprintNumber: number;
}

export interface UserStory {
  id: string;
  title: string;
  asA: string;
  iWant: string;
  soThat: string;
  acceptanceCriteria: string[];
  priority: string;
  priorityScore: number;
  tasks: TaskItem[];
}

export interface PrdContent {
  title: string;
  summary: string;
  goals: string[];
  successMetrics: string[];
  userStories: UserStory[];
}

export interface PrdDetail {
  id: string;
  title: string;
  /** @nullable */
  problem?: string | null;
  /** @nullable */
  audience?: string | null;
  /** @nullable */
  successCriteria?: string | null;
  /** @nullable */
  productName?: string | null;
  shareToken: string;
  createdAt: string;
  content: PrdContent;
}

export type PrdSummaryPriorityMix = {
  P1: number;
  P2: number;
  P3: number;
};

export interface PrdSummary {
  id: string;
  title: string;
  /** @nullable */
  problem?: string | null;
  /** @nullable */
  productName?: string | null;
  shareToken: string;
  createdAt: string;
  totalTasks: number;
  priorityMix: PrdSummaryPriorityMix;
}

export type UpdateTaskStatusBodyStatus =
  (typeof UpdateTaskStatusBodyStatus)[keyof typeof UpdateTaskStatusBodyStatus];

export const UpdateTaskStatusBodyStatus = {
  todo: "todo",
  "in-progress": "in-progress",
  done: "done",
} as const;

export interface UpdateTaskStatusBody {
  status: UpdateTaskStatusBodyStatus;
}

export interface TaskStatusRecord {
  id: string;
  prdId: string;
  taskId: string;
  status: string;
  updatedAt: string;
}

export interface PrdVersion {
  id: number;
  prdId: string;
  versionNumber: number;
  createdAt: string;
}

export interface PrdVersionDetail {
  id: number;
  prdId: string;
  versionNumber: number;
  createdAt: string;
  content: PrdContent;
}

export interface DayCount {
  date: string;
  count: number;
}

export type AdminStatsPriorityBreakdown = {
  P1: number;
  P2: number;
  P3: number;
};

export type AdminStatsEffortDistribution = {
  S: number;
  M: number;
  L: number;
  XL: number;
};

export interface AdminStats {
  totalPrds: number;
  totalTasks: number;
  avgEffortScore: number;
  prdsThisWeek: number;
  prdsPerDay: DayCount[];
  priorityBreakdown: AdminStatsPriorityBreakdown;
  effortDistribution: AdminStatsEffortDistribution;
}