# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LaunchReady API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate a new PRD
 */
export const GeneratePrdBody = zod.object({
  problem: zod.string(),
  audience: zod.string(),
  success: zod.string(),
  productName: zod.string().nullish(),
});

/**
 * @summary List all PRDs
 */
export const ListPrdsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  problem: zod.string().nullish(),
  productName: zod.string().nullish(),
  shareToken: zod.string(),
  createdAt: zod.string(),
  totalTasks: zod.number(),
  priorityMix: zod.object({
    P1: zod.number(),
    P2: zod.number(),
    P3: zod.number(),
  }),
});
export const ListPrdsResponse = zod.array(ListPrdsResponseItem);

/**
 * @summary Get a PRD by ID
 */
export const GetPrdParams = zod.object({
  id: zod.coerce.string(),
});

export const GetPrdResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  problem: zod.string().nullish(),
  audience: zod.string().nullish(),
  successCriteria: zod.string().nullish(),
  productName: zod.string().nullish(),
  shareToken: zod.string(),
  createdAt: zod.string(),
  content: zod.object({
    title: zod.string(),
    summary: zod.string(),
    goals: zod.array(zod.string()),
    successMetrics: zod.array(zod.string()),
    userStories: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        asA: zod.string(),
        iWant: zod.string(),
        soThat: zod.string(),
        acceptanceCriteria: zod.array(zod.string()),
        priority: zod.string(),
        priorityScore: zod.number(),
        tasks: zod.array(
          zod.object({
            id: zod.string(),
            title: zod.string(),
            description: zod.string(),
            dependsOn: zod.array(zod.string()),
            status: zod.string(),
            effort: zod.string(),
            effortScore: zod.number(),
            isBlocked: zod.boolean(),
            blockedBy: zod.array(zod.string()),
            sprintNumber: zod.number(),
          }),
        ),
      }),
    ),
  }),
});

/**
 * @summary Regenerate a PRD (creates new version)
 */
export const RegeneratePrdParams = zod.object({
  id: zod.coerce.string(),
});

export const RegeneratePrdResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  problem: zod.string().nullish(),
  audience: zod.string().nullish(),
  successCriteria: zod.string().nullish(),
  productName: zod.string().nullish(),
  shareToken: zod.string(),
  createdAt: zod.string(),
  content: zod.object({
    title: zod.string(),
    summary: zod.string(),
    goals: zod.array(zod.string()),
    successMetrics: zod.array(zod.string()),
    userStories: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        asA: zod.string(),
        iWant: zod.string(),
        soThat: zod.string(),
        acceptanceCriteria: zod.array(zod.string()),
        priority: zod.string(),
        priorityScore: zod.number(),
        tasks: zod.array(
          zod.object({
            id: zod.string(),
            title: zod.string(),
            description: zod.string(),
            dependsOn: zod.array(zod.string()),
            status: zod.string(),
            effort: zod.string(),
            effortScore: zod.number(),
            isBlocked: zod.boolean(),
            blockedBy: zod.array(zod.string()),
            sprintNumber: zod.number(),
          }),
        ),
      }),
    ),
  }),
});

/**
 * @summary Update task status
 */
export const UpdateTaskStatusParams = zod.object({
  id: zod.coerce.string(),
  taskId: zod.coerce.string(),
});

export const UpdateTaskStatusBody = zod.object({
  status: zod.enum(["todo", "in-progress", "done"]),
});

export const UpdateTaskStatusResponse = zod.object({
  id: zod.string(),
  prdId: zod.string(),
  taskId: zod.string(),
  status: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a PRD and all its data
 */
export const DeletePrdParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get version history for a PRD
 */
export const GetPrdVersionsParams = zod.object({
  id: zod.coerce.string(),
});

export const GetPrdVersionsResponseItem = zod.object({
  id: zod.number(),
  prdId: zod.string(),
  versionNumber: zod.number(),
  createdAt: zod.string(),
});
export const GetPrdVersionsResponse = zod.array(GetPrdVersionsResponseItem);

/**
 * @summary Get a specific version snapshot with full content
 */
export const GetPrdVersionParams = zod.object({
  id: zod.coerce.string(),
  versionId: zod.coerce.number(),
});

/**
 * @summary Get a PRD by share token (public read-only)
 */
export const GetSharedPrdParams = zod.object({
  token: zod.coerce.string(),
});

export const GetSharedPrdResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  problem: zod.string().nullish(),
  audience: zod.string().nullish(),
  successCriteria: zod.string().nullish(),
  productName: zod.string().nullish(),
  shareToken: zod.string(),
  createdAt: zod.string(),
  content: zod.object({
    title: zod.string(),
    summary: zod.string(),
    goals: zod.array(zod.string()),
    successMetrics: zod.array(zod.string()),
    userStories: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        asA: zod.string(),
        iWant: zod.string(),
        soThat: zod.string(),
        acceptanceCriteria: zod.array(zod.string()),
        priority: zod.string(),
        priorityScore: zod.number(),
        tasks: zod.array(
          zod.object({
            id: zod.string(),
            title: zod.string(),
            description: zod.string(),
            dependsOn: zod.array(zod.string()),
            status: zod.string(),
            effort: zod.string(),
            effortScore: zod.number(),
            isBlocked: zod.boolean(),
            blockedBy: zod.array(zod.string()),
            sprintNumber: zod.number(),
          }),
        ),
      }),
    ),
  }),
});

/**
 * @summary Get admin dashboard statistics
 */
export const GetAdminStatsResponse = zod.object({
  totalPrds: zod.number(),
  totalTasks: zod.number(),
  avgEffortScore: zod.number(),
  prdsThisWeek: zod.number(),
  prdsPerDay: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
  priorityBreakdown: zod.object({
    P1: zod.number(),
    P2: zod.number(),
    P3: zod.number(),
  }),
  effortDistribution: zod.object({
    S: zod.number(),
    M: zod.number(),
    L: zod.number(),
    XL: zod.number(),
  }),
});